moduls/requirements.txt
config==0.5.1
lightrag_hku==1.3.0
numpy==2.2.4
transformers==4.51.0
